feign y kafka

+++ ms-notificaciones/src/main/java/cl/teatromoro/notificaciones/client/UsuarioClient.java
@@ -8,5 +8,5 @@ import org.springframework.web.bind.annotation.PathVariable;
 public interface UsuarioClient {
 
     @GetMapping("/usuarios/{id}")
-    Object obtenerUsuario(@PathVariable("id") Long id);
+    Object obtenerUsuario(@PathVariable("id") Integer id);
 }
